rsl: refactoring
