+++ modules/paxforecast/include/motis/paxforecast/monitoring_update.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "motis/module/message.h"
+
+#include "motis/paxmon/paxmon_data.h"
+
+namespace motis::paxforecast {
+
+struct paxforecast;
+
+void on_monitoring_update(paxforecast& mod, motis::paxmon::paxmon_data& data,
+                          motis::module::msg_ptr const& msg);
+
+}  // namespace motis::paxforecast
