+++ modules/paxforecast/include/motis/paxforecast/paxforecast.h
@@ -31,10 +31,6 @@ struct paxforecast : public motis::module::module {
 
   void init(motis::module::registry&) override;
 
-private:
-  void on_monitoring_event(motis::module::msg_ptr const& msg);
-  motis::module::msg_ptr apply_measures(motis::module::msg_ptr const& msg);
-
   std::string forecast_filename_;
   std::ofstream forecast_file_;
